feat: Refactor to modules
